+++ src/test/scala/scenario/birch/Compound.scala
@@ -1,0 +1,39 @@
+package scenario.birch
+
+import draco._
+import scenario._
+import io.circe.{Decoder, Encoder, Json}
+import io.circe.syntax.EncoderOps
+
+trait Compound extends Primal[String]
+
+object Compound extends App with DracoType {
+  override lazy val typeDefinition: TypeDefinition = TypeLoader.loadType(TypeName ("Compound", _namePackage = Seq ("scenario", "birch")))
+  lazy val dracoType: Type[Compound] = Type[Compound] (typeDefinition)
+  lazy val domainType: Domain[Birch] = Domain[Birch] (typeDefinition)
+
+  implicit lazy val encoder: Encoder[Compound] = Encoder.instance { x =>
+    val fields = Seq(
+      Some("value" -> x.value.asJson)
+    ).flatten
+    Json.obj(fields: _*)
+  }
+  implicit lazy val decoder: Decoder[Compound] = Decoder.instance { cursor =>
+    for {
+      _value <- cursor.downField("value").as[Option[String]].map(_.getOrElse(""))
+    } yield Compound (_value)
+  }
+
+  def apply (
+    _value: String
+  ) : Compound = new Compound {
+    override lazy val value: String = _value
+    override lazy val typeDefinition: TypeDefinition = Compound.typeDefinition
+  }
+
+  lazy val Null: Compound = apply(
+    _value = ""
+  )
+
+
+}